linter + prettier
