+++ src/App.tsx
@@ -4,13 +4,13 @@ import "./App.css";
 import { LoginPage } from "./app/Login/Login";
 
 function App() {
-	return (
-		<div className="App">
-			<Typography variant="h1" color="textSecondary">
-				<LoginPage></LoginPage>
-			</Typography>
-		</div>
-	);
+  return (
+    <div className="App">
+      <Typography variant="h1" color="textSecondary">
+        <LoginPage></LoginPage>
+      </Typography>
+    </div>
+  );
 }
 
 export default App;
